feat: add `@deno/shim-prompts` package
