+++ README.md
@@ -5,3 +5,7 @@ Deno shims for Node.js
 ## Packages
 
 - [deno.ns](packages/deno.ns) - Deno namespace shim.
+- [@deno/shim-prompts](packages/shim-prompts) - Shims for `alert`, `confirm` and
+  `prompt`.
+- [@deno/shim-timers](packages/shim-timers) - Shims for `setTimeout` and
+  `setInterval`.
