+++ packages/deno.ns/src/index.ts
@@ -1,6 +1,5 @@
 export { webcrypto as crypto } from "crypto";
 export * as Deno from "./deno.js";
-export * from "./util/mod.js";
 
 // TODO: Wait for node builtin WhatWG Streams or fetch-blob to export them without assigning to globalThis
 import * as undici from "undici";
