+++ packages/shim-prompts/README.md
@@ -1,0 +1,3 @@
+# @deno/shim-prompts
+
+Shims for `alert`, `confirm`, and `prompt` that work in Node.js.
