+++ packages/shim-prompts/tsconfig.json
@@ -3,14 +3,16 @@
     "module": "CommonJS",
     "strict": true,
     "target": "ES2019",
-    "lib": ["ES2019"],
     "declaration": true,
     "outDir": "dist",
     "stripInternal": true
   },
   "ts-node": {
     "transpileOnly": true,
-    "files": true
+    "files": true,
+    "compilerOptions": {
+      "module": "ES2015"
+    }
   },
   "files": ["src/index.ts"]
 }
